refactor: move LDS-aware num_stages picker to gemm_config_utils
