+++ aiter/ops/triton/moe/moe_op_gemm_a4w4.py
@@ -10,10 +10,8 @@ from aiter.ops.triton._triton_kernels.moe.moe_op_gemm_a4w4 import (
     _moe_gemm_a4w4,
 )
 from aiter.ops.triton.moe.reduce import reduce_grouped
-from aiter.ops.triton.utils._triton.arch_info import (
-    get_arch,
-    pick_gemm_num_stages,
-)
+from aiter.ops.triton.utils.gemm_config_utils import pick_gemm_num_stages
+from aiter.ops.triton.utils._triton.arch_info import get_arch
 
 # -----------------------------------------------------------------------------
 #                    Matrix Multiplication + Outer Gather/Scatter
